// src/generated/models/ServiceBusModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ServiceBusEntity {
  // The entity name
  Name?: string;
  // The display name for the entity
  DisplayName?: string;
}

export interface ServiceBusMessage {
  // Content of the message
  ContentData?: string;
  // Content type of the message content
  ContentType?: string;
  // Key-value pairs for each brokered property
  Properties?: Record<string, unknown>;
  // This is a user-defined value that Service Bus can use to identify duplicate messages, if enabled.
  MessageId?: string;
  // Send to address
  To?: string;
  // Address of the queue to reply to
  ReplyTo?: string;
  // Identifier of the session to reply to
  ReplyToSessionId?: string;
  // Application specific label
  Label?: string;
  // Date and time, in UTC, when the message will be added to the queue
  ScheduledEnqueueTimeUtc?: string;
  // Identifier of the session
  SessionId?: string;
  // Identifier of the correlation
  CorrelationId?: string;
  // Identifier of the sequence number
  SequenceNumber?: number;
  // The lock token of the message as a string.
  LockToken?: string;
  // This is the duration, in ticks, that a message is valid.  The duration starts from when the message is sent to the Service Bus.
  TimeToLive?: string;
}

export interface SubscriptionFilter {
  CorrelationFilter?: SubscriptionCorrelationFilter;
}

export interface SubscriptionCorrelationFilter {
  // Identifier of the correlation
  CorrelationId?: string;
  // Application specific label
  Label?: string;
  // This is a user-defined value that Service Bus can use to identify duplicate messages, if enabled.
  MessageId?: string;
  // Key-value pairs for each brokered property
  Properties?: Record<string, unknown>;
  // Address of the queue to reply to
  ReplyTo?: string;
  // Identifier of the session to reply to
  ReplyToSessionId?: string;
  // Identifier of the session
  SessionId?: string;
  // Send to address
  To?: string;
  // Content type of the message content
  ContentType?: string;
}

export interface Subscription {
  // Subscription name.
  SubscriptionName?: string;
}

export type Object = object;